% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{boxplots}
\alias{boxplots}
\title{Create boxplots from multiple experiment}
\usage{
boxplots(df)
}
\arguments{
\item{df}{A data.frame from multiple experiment.}
}
\value{
A grid of 4 boxplots for accuracy, recall, precision and f1.
}
\description{
Create boxplots from multiple experiment
}
