% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{boxplots_coefficients}
\alias{boxplots_coefficients}
\title{Create boxplots of model coefficients from multiple experiments on artificial data}
\usage{
boxplots_coefficients(betas, true_betas)
}
\arguments{
\item{betas}{A data.frame from multiple experiment.}

\item{true_betas}{A data.frame with true values of coefficients.}
}
\value{
A boxplot presenting model coefficients.
}
\description{
Create boxplots of model coefficients from multiple experiments on artificial data
}
