% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{boxplots_influence_interactions}
\alias{boxplots_influence_interactions}
\title{Create boxplots of model accuracy scores for investigating the influence of adding the interactions to the model.}
\usage{
boxplots_influence_interactions(df)
}
\arguments{
\item{df}{A data.frame from multiple experiment.}
}
\value{
A boxplot presenting model coefficients.
}
\description{
Create boxplots of model accuracy scores for investigating the influence of adding the interactions to the model.
}
