% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_functions.R
\name{compare_models}
\alias{compare_models}
\title{A function used to compare different models to our method.}
\usage{
compare_models(train, test, tol = 1e-04, interactions = NULL, verbose = FALSE)
}
\arguments{
\item{train}{A data.frame object being the training subset with target column
as the last one.}

\item{test}{A data.frame object being the testing subset with target column
as the last one.}

\item{interactions}{A numeric vector introducing the interactions for the model.
Default value: NULL.}
}
\value{
A data.frame object with the models names and metrics values.
}
\description{
The function trains 6 different models:
\describe{
  \item{\code{DS23IRLS::irls}}{Our irls model without interactions.}
  \item{\code{DS23IRLS::irls}}{Our irls model with interactions.}
  \item{\code{msme::irls}}{The irls model from msme package.}
  \item{\code{MASS::lda}}{The lda model from MASS package.}
  \item{\code{MASS::qda}}{The qda model from MASS package.}
  \item{\code{class::knn}}{The knn model from class package.}
}
And later evaluates the with 4 metrics: accuracy, recall, precision, and f1.
}
