% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{evaluate}
\alias{evaluate}
\title{Evaluated the model based on the observed and predicted values.}
\usage{
evaluate(predictions, observed)
}
\arguments{
\item{predictions}{A numeric vector with predictions.}

\item{observed}{A numeric vector with osberved values.}
}
\value{
A list with numeric values describing the accuracy, recall, precision,
and f1 scores.
}
\description{
Evaluated the model based on the observed and predicted values.
}
\examples{
observed    <- c(1, 1, 1, 1, 0, 0 , 0, 0)
predictions <- c(0, 1, 0, 1, 0, 1 , 0, 0)
evaluate(predictions, observed)
}
