% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irls.R
\name{irls}
\alias{irls}
\title{A training function for the IRLS - Iterative Reweighted Least Squares algorithm}
\usage{
irls(X, y, tol = 1e-04, max_iter = 100, interactions = NULL)
}
\arguments{
\item{X}{A numeric data.frame object used for the fitting of the method.}

\item{y}{A numeric vector describing the target column.}

\item{max_iter}{An integer describing the maximum iteration number for the
algorithm. Default value: 100.}

\item{interactions}{A numeric vector introducing the interactions for the model.
Default value: NULL.}
}
\value{

}
\description{
A training function for the IRLS - Iterative Reweighted Least Squares algorithm
}
\examples{
X1 <- data.frame(x1 = rnorm(100), x2 = rnorm(100, 2, 5), x3 = rnorm(100, 4))
y1 <- as.matrix(rep(0, 100))
X2 <- data.frame(x1 = rnorm(100, 1, 2), x2 = rnorm(100, 12, 5), x3 = rnorm(100, 14))
y2 <- as.matrix(rep(1, 100))
X  <- rbind(X1, X2)
y  <- rbind(y1, y2)
m  <- irls(X, y)
}
