% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Calculate log likelihood value}
\usage{
log_likelihood(p, y)
}
\arguments{
\item{p}{Probability for positive class.}

\item{y}{True labels.}
}
\value{
Log-likelihood value.
}
\description{
Calculate log likelihood value
}
