% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_functions.R
\name{multiple_compare_interactions}
\alias{multiple_compare_interactions}
\title{Runs a multiple experiment on artificial data in order to check the influence of adding interactions to model on model accuracy.}
\usage{
multiple_compare_interactions(ds, interactions, L = 50, verbose = FALSE)
}
\arguments{
\item{ds}{A function generating a dataset, the target variable should be in column 'y'.}

\item{interactions}{A dataframe in which rows define the interactions.}

\item{L}{Number of repetitions.}

\item{verbose}{A logical value indicating whether we want to print the string
or not.}
}
\value{
A dataframe containing the accuracy scores for models with interactions and without interactions.
}
\description{
Runs a multiple experiment on artificial data in order to check the influence of adding interactions to model on model accuracy.
}
