% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_functions.R
\name{multiple_exp}
\alias{multiple_exp}
\title{Runs a multiple experiment on different train and test subset sizes.}
\usage{
multiple_exp(
  p = c(0.9, 0.8, 0.7, 0.6, 0.5),
  include_occupancy = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{p}{A numeric vector, where every value is between 0 and 1. Describes
the training set size (the testing is 1 - p).}

\item{include_occupancy}{A logical value, describing if we should include
computations for a time consuming occupancy dataset.}

\item{verbose}{A logical value indicating whether we want to print the string
or not.}
}
\value{
A list of data frames with scores from multiple expermients.
}
\description{
Runs a multiple experiment on different train and test subset sizes.
}
