% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_functions.R
\name{multiple_not_converge}
\alias{multiple_not_converge}
\title{Runs a multiple experiment on artificial data in order to check if method converges.}
\usage{
multiple_not_converge(ds, L = 50, verbose = FALSE)
}
\arguments{
\item{ds}{A function generating a dataset, the features should be in columns 'x1', 'x2', ..., and the target variable should be in column 'y'}

\item{L}{Number of repetitions.}

\item{verbose}{A logical value indicating whether we want to print the string
or not.}
}
\value{
A percentage value of the experiments in which the method did not converge.
}
\description{
Runs a multiple experiment on artificial data in order to check if method converges.
}
