% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_occupancy.R
\docType{data}
\name{occupancy}
\alias{occupancy}
\title{Occupancy dataset}
\format{

}
\source{
Data from UCL \url{https://archive.ics.uci.edu/ml/datasets/Occupancy+Detection+}
}
\usage{
data(occupancy)
}
\description{
Experimental data used for binary classification (room occupancy) from
Temperature,Humidity,Light and CO2. Ground-truth occupancy was obtained
from time stamped pictures that were taken every minute. We've removed the
date time column of the measurement as it is unnecessary for the training model.
}
