% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_heatmaps}
\alias{plot_heatmaps}
\title{Creates plots for the correlation heatmaps}
\usage{
plot_heatmaps(cor_raisin, cor_occupancy, cor_banknote)
}
\arguments{
\item{cor_raisin}{The NxN matrix, describing the correlation matrix for the
raisin dataset.}

\item{cor_occupancy}{The NxN matrix, describing the correlation matrix for
the occupancy dataset.}

\item{cor_banknote}{The NxN matrix, describing the correlation matrix for
the banknote dataset.}
}
\description{
Creates plots for the correlation heatmaps
}
