% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.irls}
\alias{predict.irls}
\title{Predict function for irls model}
\usage{
\method{predict}{irls}(object, X, prob = TRUE, ...)
}
\arguments{
\item{object}{An irls object from `irls()` function.}

\item{X}{A data.frame object with a dataset to predict.}

\item{prob}{A logical value determining if we want probabilities values in the output.}

\item{...}{Further arguments passed to or from other methods (currently ignored).}
}
\value{
A data.frame with predictions (or their probabilities).
}
\description{
Predict function for irls model
}
