% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_raisin.R
\docType{data}
\name{raisin}
\alias{raisin}
\title{Raisin dataset}
\format{
A data frame with 900 rows and 8 variables:
\describe{
  \item{Area}{integer, Gives the number of pixels within the boundaries of the raisin.}
  \item{Perimeter}{numeric, It measures the environment by calculating the distance between the boundaries of the raisin and the pixels around it.}
  \item{MajorAxisLength}{numeric, Gives the length of the main axis, which is the longest line that can be drawn on the raisin.}
  \item{MinorAxisLength}{numeric, Gives the length of the small axis, which is the shortest line that can be drawn on the raisin.}
  \item{Eccentricity}{numeric, It gives a measure of the eccentricity of the ellipse, which has the same moments as raisins.}
  \item{ConvexArea}{integer, Gives the number of pixels of the smallest convex shell of the region formed by the raisin.}
  \item{Extent}{numeric, Gives the ratio of the region formed by the raisin to the total pixels in the bounding box.}
  \item{Class}{integer, Kecimen and Besni raisin.}
}
}
\source{
Data from UCL \url{https://archive.ics.uci.edu/ml/datasets/Raisin+Dataset}
}
\usage{
data(raisin)
}
\description{
Images of Kecimen and Besni raisin varieties grown in Turkey were obtained
with CVS. A total of 900 raisin grains were used, including 450 pieces from
both varieties. These images were subjected to various stages of pre-processing
and 7 morphological features were extracted. These features have been classified
using three different artificial intelligence techniques.
}
