% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.irls}
\alias{summary.irls}
\title{Describes an irls object briefly}
\usage{
\method{summary}{irls}(object, ...)
}
\arguments{
\item{object}{An irls object from `irls()` function.}

\item{...}{Further arguments passed to or from other methods (currently ignored).}
}
\value{
A brief description of the irls object.
}
\description{
Describes an irls object briefly
}
\examples{
X1 <- data.frame(x1 = rnorm(100), x2 = rnorm(100, 2, 5), x3 = rnorm(100, 4))
y1 <- as.matrix(rep(0, 100))
X2 <- data.frame(x1 = rnorm(100, 1, 2), x2 = rnorm(100, 12, 5), x3 = rnorm(100, 14))
y2 <- as.matrix(rep(1, 100))
X  <- rbind(X1, X2)
y  <- rbind(y1, y2)
m  <- irls(X, y)
summary(m)
}
