% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{train_test_split}
\alias{train_test_split}
\title{Splits dataset into train, and test subsets}
\usage{
train_test_split(df, p = c(0.8, 0.2), seed = NULL)
}
\arguments{
\item{df}{A data.frame object, the data set you want to split.}

\item{p}{A numeric vector of length 2 which variables sum to 1 indicating
the proportions for train and test subsets. Default value: c(0.8, 0.2).}

\item{seed}{An integer describing the seed for the splitting algorithm.
Default value NULL.}
}
\value{
A list of two data.frames for train and test subsets.
}
\description{
Splits dataset into train, and test subsets
}
\examples{
train_test_split(banknote)
}
