% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irls.R
\name{irls}
\alias{irls}
\title{A training function for the IRLS - Iterative Reweighted Least Squares algorithm}
\usage{
irls(X, y, tol = 1e-04, max_iter = 100, interactions = NULL, verbose = FALSE)
}
\arguments{
\item{X}{A numeric data.frame object used for the fitting of the method.}

\item{y}{A numeric vector describing the target column.}

\item{tol}{A difference between log-likelihood function values between two iterations below which the algorithm should stop.}

\item{max_iter}{An integer describing the maximum iteration number for the
algorithm. Default value: 100.}

\item{interactions}{A numeric vector introducing the interactions for the model.
Default value: NULL.}

\item{verbose}{A logical value indicating whether we want to have prints or not.}
}
\value{
An object of class irls containing the objects:
\item{\code{tol}}{}
\item{\code{max_iter}}{An integer describing the maximum iteration number for the
algorithm.}
\item{\code{n_iter}}{An integer describing the actual number of iterations for the
algorithm.}
\item{\code{beta}}{A numeric matrix with coefficient values for the data set columns.}
\item{\code{colnames}}{A character vector with column names from the data frame.}
\item{\code{interactions}}{A numeric vector describing the interactions for the model.}
}
\description{
A training function for the IRLS - Iterative Reweighted Least Squares algorithm
}
\examples{
X1 <- data.frame(x1 = rnorm(100), x2 = rnorm(100, 2, 5), x3 = rnorm(100, 4))
y1 <- as.matrix(rep(0, 100))
X2 <- data.frame(x1 = rnorm(100, 1, 2), x2 = rnorm(100, 12, 5), x3 = rnorm(100, 14))
y2 <- as.matrix(rep(1, 100))
X  <- rbind(X1, X2)
y  <- rbind(y1, y2)
m  <- irls(X, y)
}
