% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.irls}
\alias{predict.irls}
\title{Predict function for irls model}
\usage{
\method{predict}{irls}(object, X, prob = TRUE, ...)
}
\arguments{
\item{object}{An irls object from `irls()` function.}

\item{X}{A data.frame object with a dataset to predict.}

\item{prob}{A logical value determining if we want probabilities values in the output.}

\item{...}{Further arguments passed to or from other methods (currently ignored).}
}
\value{
A data.frame with predictions (or their probabilities).
}
\description{
Predict function for irls model
}
\examples{
X1 <- data.frame(x1 = rnorm(100), x2 = rnorm(100, 2, 5), x3 = rnorm(100, 4))
y1 <- as.matrix(rep(0, 100))
X2 <- data.frame(x1 = rnorm(100, 1, 2), x2 = rnorm(100, 12, 5), x3 = rnorm(100, 14))
y2 <- as.matrix(rep(1, 100))
X  <- rbind(X1, X2)
y  <- rbind(y1, y2)
X_train <- X
m <- irls(X, y)
preds <- predict(m, X, prob = FALSE)
preds
}
